+++ astroquery/sdss/tests/test_sdss_remote.py
@@ -171,13 +171,13 @@ class TestSDSSRemote:
         assert isinstance(query1, Table)
         assert isinstance(query2, Table)
 
-        assert query1.colnames == ['r', 'psfMag_r']
-        assert query2.colnames == ['ra', 'dec', 'r']
+        assert query1.colnames == ['objID', 'r', 'psfMag_r']
+        assert query2.colnames == ['objID', 'ra', 'dec', 'r']
 
     @pytest.mark.parametrize("dr", dr_list)
     def test_query_crossid(self, dr):
         query1 = sdss.SDSS.query_crossid(self.coords, data_release=dr)
-        query2 = sdss.SDSS.query_crossid([self.coords, self.coords])
+        query2 = sdss.SDSS.query_crossid([self.coords, self.coords], data_release=dr)
         assert isinstance(query1, Table)
         assert query1['objID'][0] == 1237652943176138868